% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aminoAcidsPlot.R
\name{aminoAcidsPlot}
\alias{aminoAcidsPlot}
\title{Amino Acids Plotting}
\usage{
aminoAcidsPlot(aminoAcids)
}
\arguments{
\item{aminoAcids}{Define an amino acid}
}
\value{
an amino acid plot showing the amount of each unique AA
}
\description{
Amino Acids Plotting
}
