% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_to_aa_sequence.R
\name{codons_to_aa_sequence}
\alias{codons_to_aa_sequence}
\title{Codons to Amino Acid Sequence}
\usage{
codons_to_aa_sequence(codons)
}
\arguments{
\item{codons}{a vector of 3-letter codons as strings}
}
\value{
A collapsed amino acid sequence as a string
}
\description{
Codons to Amino Acid Sequence
}
