% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_sequence.R
\name{dna_sequence}
\alias{dna_sequence}
\title{Title
Create a DNA sequence from A C G C with defined lenght and repeating values}
\usage{
dna_sequence(string_length)
}
\arguments{
\item{string_length}{Desired length of sequence (int)}
}
\value{
A DNA sequence (str(?))
}
\description{
Title
Create a DNA sequence from A C G C with defined lenght and repeating values
}
