% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_codons.R
\name{plot_codons}
\alias{plot_codons}
\title{Plot codon count, using}
\usage{
plot_codons(codon_string)
}
\arguments{
\item{codon_string}{A string of codons (one letter), you want to plot}
}
\value{
Returns an object containing a plot
}
\description{
Plot codon count, using
}
\examples{
codon_string = "MRTYVCIVC"
plot_codons(codon_string)
}
