% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinTraduction.R
\name{proteinTraduction}
\alias{proteinTraduction}
\title{Protein Traduction}
\usage{
proteinTraduction(codons)
}
\arguments{
\item{codons}{Strings, provide a vector with codons to traduce to AA}
}
\value{
String, the protein formed by the created AA
}
\description{
Protein Traduction
}
