% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomDna.R
\name{randomDna}
\alias{randomDna}
\title{Rnadom DNA Sample}
\usage{
randomDna(inputSize)
}
\arguments{
\item{inputSize}{Int, input size for desired DNA random sample}
}
\value{
String, outputs a DNA random sample based on the provided size
}
\description{
Rnadom DNA Sample
}
