% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_T_with_U.R
\name{replace_T_with_U}
\alias{replace_T_with_U}
\title{DNA to RNA}
\usage{
replace_T_with_U(replace_T_with_U_2)
}
\arguments{
\item{dna_str}{A string vector in which you want to replace 'T' with 'U'.}
}
\value{
rna_str A string vector with all 'T's replaced by 'U's.
}
\description{
This function replaces all of 'T' with 'U' in a given DNA chain vector.
}
\examples{
replace_T_with_U("TACGT")  # Returns "UACGU"
replace_T_with_U("TTTTT")  # Returns "UUUUU"
replace_T_with_U("AGCTA")  # Returns "AGCUA"

}
