% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceToCodons.R
\name{sequenceToCodons}
\alias{sequenceToCodons}
\title{Title}
\usage{
sequenceToCodons(sequence, start = 1)
}
\arguments{
\item{sequence}{the DNA sequence}

\item{start}{starting position}
}
\value{
codons from the sequence
}
\description{
Title
}
