% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codons.R
\name{get_codons}
\alias{get_codons}
\title{Get codons}
\usage{
get_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{Instert string of a DNA or RNA sequence}

\item{start}{The first base pair in the sequence to be included}
}
\value{
A vector of strings of codons in the sequence
}
\description{
Get codons
}
